+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Allow to disable WrapperWidgetUI [#420](https://github.com/CartoDB/carto-react/pull/420)
+
 ## 1.3
 
 ### 1.3.0-alpha.12 (2022-06-01)